+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- [client] Unmanaged proxies (when using `native_lib`) are now correctly considered alive rather than dead
+
 ## 0.21.1 - 2018-09-25
 
 - [server] When using `native_lib`, only attempt to filter globals that are managed by the rust lib.